moves enable coupons call to coupons controller
